androidmedia: Support float i-frame-interval
Android 25 added support for i-frame-interval to be a floating
point value. Store the property as a float and use the newer
version when it's available.

+++ sys/androidmedia/gstamcvideoenc.h
@@ -63,7 +63,7 @@ struct _GstAmcVideoEnc
   GstAmcColorFormatInfo color_format_info;
 
   guint bitrate;
-  guint i_frame_int;
+  gfloat i_frame_int;
 
   /* TRUE if the component is configured and saw
    * the first buffer */
